Add checking if user exists in database
